add some more things to the algorithm..

+++ src/data_prep/mnist_data.py
@@ -19,7 +19,7 @@ def fetch_mnistdata(train_pct, classes):
     '''
     mnist = fetch_mldata('MNIST original')
     data = mnist.data
-    target = mnist.target
+    target = mnist.target.astype(int) # convert labels to int
 
     data = data[np.logical_or.reduce([target == cls for cls in classes])]  
     target = target[np.logical_or.reduce([target == cls for cls in classes])]
